requirements.txt
asgiref==3.9.1
beautifulsoup4==4.13.4
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
colorama==0.4.6
Django==4.2.20
django-axes==8.0.0
django-cleanup==9.0.0
django-debug-toolbar==5.1.0
django-filter==23.1
django-ratelimit==4.1.0
djangorestframework==3.16.1
djangorestframework-simplejwt==4.7.2
djoser==2.3.3
drf-yasg==1.21.10
et_xmlfile==2.0.0
flake8==7.3.0
flake8-isort==6.1.2
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
isort==6.0.1
lxml==6.0.0
mccabe==0.7.0
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pillow==11.3.0
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pyflakes==3.4.0
pyTelegramBotAPI==4.29.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
setuptools==80.9.0
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
yandex_tracker_client==2.9
